03_activity/requirements.txt
mysql-connector-python
chardet
pyspark